api/serializers.py
from rest_framework import serializers
from .models import Category, Tag, Item
from django.contrib.auth.models import User

class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'

class TagSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tag
        fields = '__all__'

class ItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Item
        fields = '__all__'

class UserSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ('username', 'password')

    def create(self, validated_data):
        user = User.objects.create_user(**validated_data)
        return user

class ItemCreateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Item
        fields = ('sku', 'name', 'category', 'tags', 'stock_status', 'available_stock')